Use mdpopups to display docstring popup when hovering text.
If available the first line is highlighted using mdpopups highlighting feature. The rest of the text is handled as markdown. Finally some maybe common headlines are highlighted.

If mdpopups is not available, rendering falls back to simple html creation.

If "enable_tooltip" is set to true in jedi's settings, the default goto definition tooltip is disabled for python sources in order to avoid fighting.