Fix issue with multiple line signatures in docstring tooltips
The signature in a docstring may span multiple lines. Don't know whether this is a Jedi bug, but we need to handle it.

Therefore the regex pattern is used to extract the signature from the whole docstring instead of assuming the signature to be part of the first line only.