Increase docstring signature font-size and add tooltip styling guide
Tooltips if displayed using mdpopups can be modified using mdpopups.css file using a `.jedi` selector. This commit adds the description how to do so and where to find further documantation to the README.

The default font size for the code signature in docstring tooltips is increased to make it stand out more clearly.